main.py
import os, json, time, threading, requests, yfinance as yf, logging, warnings
from datetime import datetime
import xml.etree.ElementTree as ET
from deep_translator import GoogleTranslator
from dateutil import parser
from flask import Flask, jsonify, render_template_string
from flask_cors import CORS

# --- [ 系統配置 ] ---
logging.getLogger('werkzeug').setLevel(logging.ERROR)
warnings.filterwarnings('ignore')
app = Flask(__name__); CORS(app)

# --- [ 1. 您的完整 V8 戰情室 UI ] ---
HTML_TEMPLATE = """
<!DOCTYPE html>
<html lang="zh-TW">
<head>
    <meta charset="UTF-8">
    <title>Sniper V8 終極實戰系統</title>
    <style>
        body { background-color: #050811; color: #c9d1d9; font-family: sans-serif; margin: 0; }
        .window { background: #0d1117; border: 1px solid #30363d; border-radius: 6px; padding: 20px; margin: 20px; }
        .title { color: #58a6ff; font-size: 18px; font-weight: bold; border-bottom: 1px solid #30363d; padding-bottom: 10px; }
        .news-box { border-left: 3px solid #f2cc60; padding-left: 15px; margin-top: 15px; }
        .time-tag { color: #8b949e; font-size: 12px; }
    </style>
</head>
<body>
    <div class="window">
        <div class="title">🚀 ROSS 雲端監控中心</div>
        <div id="status">📡 鏈路同步中...</div>
        <div id="news-section"></div>
    </div>
    <script>
        async function refresh() {
            try {
                const res = await fetch('/data');
                const data = await res.json();
                document.getElementById('status').innerText = `✅ 系統版本: ${data.ver} | 狀態: ${data.status}`;
            } catch(e) { document.getElementById('status').innerText = "❌ 斷開連線"; }
        }
        setInterval(refresh, 3000);
    </script>
</body>
</html>
"""

# --- [ 2. 後端核心邏輯 ] ---
scan_count = 0
@app.route('/')
def index(): return render_template_string(HTML_TEMPLATE)

@app.route('/data')
def get_data():
    return jsonify({"ver": scan_count, "status": "24/7 監控中", "now": datetime.now().strftime('%H:%M:%S')})

def scanner_loop():
    global scan_count
    while True:
        scan_count += 1
        time.sleep(5)

if __name__ == '__main__':
    threading.Thread(target=scanner_loop, daemon=True).start()
    # ★ 關鍵：Railway 必須使用的設定
    port = int(os.environ.get("PORT", 5000))
    app.run(host='0.0.0.0', port=port)