requirements.txt
flask
flask-cors
yfinance
deep-translator
python-dateutil
lxml
beautifulsoup4
requests
pytz